dataprovider/requirements.txt
fastapi
uvicorn
pymongo
python-jose
pydantic
